requirements.txt
# Core LangChain packages - updated for v0.3
langchain==0.3.26
langchain-core>=0.3.66,<1.0.0
langchain-community==0.3.27

# LangChain integrations
langchain-google-genai==2.1.7
langchain-groq==0.3.6
langchain-huggingface==0.3.0
langchain-text-splitters>=0.3.0

# Pydantic 2 - required for LangChain v0.3 (Pydantic 1 no longer supported)
pydantic>=2.0.0,<3.0.0

# Web crawling and data processing
crawl4ai==0.7.0
requests==2.32.4
pandas==2.3.1
beautifulsoup4~=4.12
lxml==5.3.2

# Vector database
weaviate-client==4.16.0

# ML and AI models
huggingface-hub==0.33.4
scikit-learn==1.6.1
torch==2.7.1

# Environment management
python-dotenv~=1.0

# OCR file extraction
pymupdf==1.26.3
pytesseract>=0.3.10
